Changed over to mod-version 4.0
Addedd in preview.

git diff HEAD

Modified 3.0 versions.

Added in 3.0 differences.

Apparently the new default cmake on macos has removed compatibily with this cmake, so let's see if this fixes it.

Ensured that --reinstall also reinstalls lite-xls.

Added in the capability of specifying a bottle's location, and constructing them as folders outside the normal flow.

Ensured version comes out in JSON.

Removed simplfied.

Sigh.

Ensured that we only try to copy the system config when it actually exists.